Move alacarte to examples
